+++ src/manager/agents/common-module/module/utils/common_utils.hpp
@@ -15,30 +15,30 @@ class CommonUtils
 {
 public:
   static std::map<std::string, ScAddr> managerParametersWithAgentRelations;
+  static std::vector<std::vector<ScAddr>> componentsClasses;
 
   static void InitParametersMap();
-
   static bool TransformToScStruct(
-      ScMemoryContext & m_memoryCtx,
+      ScMemoryContext & context,
       ScAddr const & actionAddr,
       std::map<std::string, std::vector<std::string>> const & commandParameters);
-
+  static ScAddr GetMyselfDecompositionAddr(ScMemoryContext & context);
+  static void CreateMyselfDecomposition(ScMemoryContext & context);
   static ScAddrVector GetNodesUnderParameter(
-      ScMemoryContext & m_memoryCtx,
+      ScMemoryContext & context,
       ScAddr const & actionAddr,
       ScAddr const & relationAddr);
-
   static ScAddr GetParameterNodeUnderRelation(
-      ScMemoryContext & m_memoryCtx,
+      ScMemoryContext & context,
       ScAddr const & actionAddr,
       ScAddr const & relation);
-
-  static std::map<std::string, ScAddr> GetSetElements(ScMemoryContext & m_memoryCtx, ScAddr const & setAddr);
-
-  static std::map<std::string, ScAddr> GetElementsLinksOfSet(ScMemoryContext & m_memoryCtx, ScAddr const & setAddr);
-
+  static std::map<std::string, ScAddr> GetSetElements(ScMemoryContext & context, ScAddr const & setAddr);
+  static std::map<std::string, ScAddr> GetElementsLinksOfSet(ScMemoryContext & context, ScAddr const & setAddr);
   static std::map<std::string, std::vector<std::string>> GetCommandParameters(
-      ScMemoryContext & m_memoryCtx,
+      ScMemoryContext & context,
       ScAddr const & actionAddr);
+  static ScAddr GetSubsystemDecompositionAddr(ScMemoryContext & context, ScAddr const & component);
+  static bool CheckIfInstalled(ScMemoryContext & context, ScAddr const & component);
+  static ScAddr GetComponentBySpecification(ScMemoryContext & context, ScAddr const & specification);
 };
 }  // namespace common_utils